% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cards}
\alias{cards}
\title{Card details}
\format{
A data frame with 228 rows and 18
variables:
\subsection{Names}{

\describe{
\item{name}{Name of card}
\item{version}{Version of card. Some cards share their \code{name}, The version,
helps make distinctions between them. Will sometimes be \code{NA}.}
}
}

\subsection{Gameplay}{

\describe{
\item{card_type}{Card type, takes values \code{Action}, \verb{Action, Song},
\code{Character}, or \code{Item}.}
\item{classifications}{Comma separated, types of classes. Such as \code{Ally},
\code{King} and \code{Villain}.}
\item{ink_color}{Ink color, takes values \code{Amber}, \code{Amethyst}, \code{Emerald},
\code{Ruby}, \code{Sapphire}, or \code{Steel}.}
\item{ink_cost}{Integer value, denoting its ink cost.}
\item{inkwell}{Logical value, denoting if card can be put in inkwell or
not. Takes values \code{Yes} and \code{No}.}
\item{strength}{Integer value, denoting the card's strength. Will be \code{NA}
for non-character cards.}
\item{willpower}{Integer value, denoting the card's willpower Will be \code{NA}
for non-character cards.}
\item{keywords_abilities}{Character, extracted keywords and abilities.}
\item{lore}{Integer value, amount of lore card will produce.}
}
}

\subsection{Card information}{

\describe{
\item{card_id}{Character, on the format \code{x/y} where \code{x} is the number the
card as in the set, and \code{y} is the number of of cards in the set. \code{x} can
be larger than \code{y}.}
\item{card_text}{Character, text of the card. Can be \code{NA}.}
\item{flavor_text}{Character, text of the card. Can be \code{NA}.}
\item{set}{Character, name of set card is printed in.}
\item{rarity}{Character, rarity of card. Takes values: \code{Common},
\code{Uncommon}, \code{Rare}, \verb{Super Rare}, \code{Promo}, \code{Legendary}, \code{Enchanted}.}
\item{illustrator}{Character, name of illustrator.}
\item{release_date}{Character, release date.}
}
}
}
\source{
\url{https://lorcanaplayer.com/}

\url{https://lorcania.com/}
}
\usage{
cards
}
\description{
A data set containing all the details on the lorcana cards.
}
\keyword{datasets}
